+++ CHANGELOG.md
@@ -82,6 +82,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fixed internal error when updating or creating a sale with missing required values - #4778 by @NyanKiyoshi
 - Fixed the internal error filtering pages by URL in the dashboard 1.0 - #4776 by @NyanKiyoshi
 - Added product variant bulk create mutation - #4735 by @fowczarek
+- Added product variant bulk create mutation - #4749 by @fowczarek
 
 ## 2.8.0